onboarding flow refactor and new steps
