+++ src/pages/staking-hub/onboarding/components.tsx
