+++ src/pages/staking-hub/onboarding/safeOnboarding/styled.tsx
@@ -7,8 +7,8 @@ import {
 } from '@mui/material'
 
 import styled from '@emotion/styled';
-import Button from '../../future-hopr-lib-components/Button';
-import GrayButton from '../../future-hopr-lib-components/Button/gray';
+import Button from '../../../../future-hopr-lib-components/Button';
+import GrayButton from '../../../../future-hopr-lib-components/Button/gray';
 
 export const Lowercase = styled.span`
   text-transform: lowercase;
